+++ packages/connect/src/relay/state.ts
@@ -1,8 +1,12 @@
 import type { HoprConnectOptions, Stream } from '../types.js'
 import type { PeerId } from '@libp2p/interface-peer-id'
 
-import { nAtATime } from '@hoprnet/hopr-utils'
-import { RelayContext, type RelayContextInterface } from './context.js'
+import { nAtATime, u8aCompare } from '@hoprnet/hopr-utils'
+import { IStream, Server, connect_relay_set_panic_hook } from '../../lib/connect_relay.js'
+import { webcrypto } from 'node:crypto'
+// @ts-ignore
+globalThis.crypto = webcrypto
+connect_relay_set_panic_hook()
 
 import debug from 'debug'
 
@@ -12,7 +16,7 @@ const verbose = debug(DEBUG_PREFIX.concat(':verbose'))
 const error = debug(DEBUG_PREFIX.concat(':error'))
 
 type RelayConnections = {
-  [id: string]: RelayContextInterface
+  [id: string]: Server
 }
 
 /**
@@ -92,7 +96,7 @@ class RelayState {
       return false
     }
 
-    context[source.toString()].update(toSource)
+    context[source.toString()].update(toSource as IStream)
 
     return true
   }
@@ -114,7 +118,7 @@ class RelayState {
    * Performs an operation for each relay context in the current set.
    * @param action
    */
-  async forEach(action: (dst: string, ctx: RelayContextInterface) => Promise<void>) {
+  async forEach(action: (dst: string, ctx: Server) => Promise<void>) {
     await nAtATime(
       action,
       this,
@@ -155,17 +159,37 @@ class RelayState {
     toDestination: Stream,
     __relayFreeTimeout?: number
   ): Promise<void> {
-    const toSourceContext = RelayContext(
-      toSource,
+    const toSourceContext = new Server(
+      {
+        source: (async function* () {
+          for await (const maybeBuf of toSource.source) {
+            if (Buffer.isBuffer(maybeBuf)) {
+              yield new Uint8Array(maybeBuf.buffer, maybeBuf.byteOffset, maybeBuf.length)
+            } else {
+              yield maybeBuf
+            }
+          }
+        })(),
+        sink: toSource.sink
+      },
+      // toSource as IStream,
       { onClose: this.cleanListener(source, destination), onUpgrade: this.cleanListener(source, destination) },
       this.options
     )
-    const toDestinationContext = RelayContext(
-      toDestination,
+    const toDestinationContext = new Server(
       {
-        onClose: this.cleanListener(destination, source),
-        onUpgrade: this.cleanListener(destination, source)
+        source: (async function* () {
+          for await (const maybeBuf of toDestination.source) {
+            if (Buffer.isBuffer(maybeBuf)) {
+              yield new Uint8Array(maybeBuf.buffer, maybeBuf.byteOffset, maybeBuf.length)
+            } else {
+              yield maybeBuf
+            }
+          }
+        })(),
+        sink: toDestination.sink
       },
+      { onClose: this.cleanListener(destination, source), onUpgrade: this.cleanListener(destination, source) },
       this.options
     )